fix(utils):change global css, change text color , change dashboard page to root page
